prevent multiple instances of the same project file to be opened at the same time

+++ StoryEditor/NewProjectWizard.cs
@@ -384,7 +384,7 @@ namespace OneStoryProjectEditor
 				foreach (var aTeamMember in
 					_storyProjectData.TeamMembers.Values.Where(aTeamMember => aTeamMember.MemberType == TeamMemberData.UserTypes.eProjectFacilitator))
 				{
-					aTeamMember.DefaultAllowed |= (long)newTaskAllowed;
+					aTeamMember.DefaultAllowed |= (long)(uint)newTaskAllowed;
 				}
 			}
 		}
